add Job.result() across all SDKs, fix CI lint/type errors
- Add result() / result property to Job in all 5 SDKs that parses
  processedData when the job is complete, returns nil/None/null otherwise
- Python: fix unused imports (ruff), mypy type errors in upload + backoff
- TypeScript: add @types/node, DOM lib types, fix fetch body type cast
- Rust: run cargo fmt, suppress clippy::type_complexity on callback field

+++ python/src/framequery/_async_client.py
@@ -241,4 +241,4 @@ def _backoff_delay(attempt: int, response: Optional[httpx.Response] = None) -> f
                 return float(ra)
             except ValueError:
                 pass
-    return min(0.5 * (2**attempt), 30.0)
+    return float(min(0.5 * (2**attempt), 30.0))
